# server/openapi.yaml
openapi: '3.0.3'
info:
  title: ESA Dono API
  description: |
    Donation platform with Stripe checkout, wallet balance, pledge (smart cart)
    auto-fulfillment, rewards, polls, fund goals, and moderator/admin tooling.

    ## Authentication

    | Method | Mechanism | Used By |
    |--------|-----------|---------|
    | Donor token | `?token=<magic_token>` query param on every request | Donor wallet, spend endpoints, moderator |
    | Admin key | `X-Admin-Key: <ADMIN_API_KEY>` header | Admin routes; also a fallback for moderator routes |
    | Moderator key | `X-Moderator-Key: <MODERATOR_API_KEY>` header | Fallback access to moderator routes, independent of donor roles |
    | Moderator/Admin role | Donor token + `role: MODERATOR` or `role: ADMIN` on donor record | Moderator routes (ADMIN implies MODERATOR access) |
    | Stripe webhook | HMAC signature via `stripe-signature` header | Stripe webhook (skipped when `STRIPE_WEBHOOK_SECRET` unset) |

    ## Rate Limiting

    Spend endpoints (`claim`, `vote`, `custom-entry`, `contribute`) are rate-limited:
    20 requests per minute, keyed by `?token=` or IP. Configurable via `RATE_LIMIT_SPEND`.

    ## Wallet Discount

    When `POST /api/pledge` receives a valid `?token=` (authenticated donor),
    the donor's `balance_remaining` is applied as a discount on the Stripe
    checkout amount. If the wallet covers the entire pledge total, the pledge
...

    ## Money

...
  version: '1.0.0'
servers:
  - url: /api
...
paths:
...
  /admin/stats:
    get:
      tags: [Admin]
      summary: Admin dashboard stats
      security: [{ AdminKey: [] }]
      responses:
        '200':
          description: Aggregated stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  donors: { type: integer }
                  donations: { type: integer }
                  claims: { type: integer }
                  pledges: { type: integer }
                  total_raised_cents: { type: integer }
                  streams:
                    type: array
                    description: Per-stream raised totals (admin-only; not shown publicly)
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                        raised_cents: { type: integer }
                        donations: { type: integer }
  /admin/streams:
    get:
      tags: [Admin]
      summary: List all streams
      security: [{ AdminKey: [] }]
      responses:
        '200':
          description: All streams ordered by created_at ascending
    post:
      tags: [Admin]
      summary: Create a stream
      security: [{ AdminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                is_active: { type: boolean, default: true }
      responses:
        '200':
          description: Stream created
        '409':
          description: Stream name already exists
  /admin/streams/{id}:
    put:
      tags: [Admin]
      summary: Update a stream
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                is_active: { type: boolean }
      responses:
        '200':
          description: Stream updated
    delete:
      tags: [Admin]
      summary: Deactivate a stream
      description: Soft-delete — sets is_active false rather than removing the row, to preserve references from incentives/donations/pledges.
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Stream deactivated
  /admin/donations:
    get:
      tags: [Admin]
      summary: List all donations
      security: [{ AdminKey: [] }]
      responses:
        '200':
          description: Donations with donor email, ordered by created_at descending
  /admin/claims:
    get:
      tags: [Admin]
      summary: List all reward claims
      security: [{ AdminKey: [] }]
      responses:
        '200':
          description: Claims with reward and donor email
  /admin/claims/{id}:
    patch:
      tags: [Admin]
      summary: Update claim status
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [PENDING, FULFILLED]
      responses:
        '200':
          description: Claim updated
  /admin/rewards:
    get:
      tags: [Admin]
      summary: List all rewards
      security: [{ AdminKey: [] }]
      responses:
        '200':
          description: All rewards ordered by created_at descending
    post:
      tags: [Admin]
      summary: Create a reward
      security: [{ AdminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, type, cost_cents]
              properties:
                title: { type: string }
                description: { type: string }
                type:
                  type: string
                  enum: [DIGITAL, PHYSICAL, SHOUTOUT, CUSTOM]
                cost_cents: { type: integer }
                quantity_total: { type: integer, nullable: true }
                is_active: { type: boolean, default: true }
                custom_type_label: { type: string }
                stream_id: { type: string, nullable: true }
      responses:
        '200':
          description: Reward created
  /admin/rewards/{id}:
    put:
      tags: [Admin]
      summary: Update a reward
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                type: { type: string, enum: [DIGITAL, PHYSICAL, SHOUTOUT, CUSTOM] }
                cost_cents: { type: integer }
                quantity_total: { type: integer, nullable: true }
                is_active: { type: boolean }
                custom_type_label: { type: string }
                stream_id: { type: string, nullable: true }
      responses:
        '200':
          description: Reward updated
    delete:
      tags: [Admin]
      summary: Delete a reward
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
  /admin/simulate-donation:
    post:
      tags: [Admin]
      summary: Simulate a donation (no real money)
      security: [{ AdminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, amount_cents]
              properties:
                email: { type: string, format: email }
                donor_name: { type: string }
                amount_cents: { type: integer, minimum: 100 }
                comment: { type: string }
                pledge_token: { type: string }
      responses:
        '200':
          description: Donation simulated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  token: { type: string, description: 'Donor magic token' }
                  donor:
                    type: object
                    properties:
                      id: { type: string }
                      email: { type: string }
                      balance_remaining: { type: integer }
                  pledge: { type: object, nullable: true }
  /admin/donors:
    get:
      tags: [Admin]
      summary: List donors
      security: [{ AdminKey: [] }]
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Search by email
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
          description: Pagination offset
      responses:
        '200':
          description: Paginated donor list (capped at 50)
  /admin/donors/{id}:
    get:
      tags: [Admin]
      summary: Get donor details
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Full donor with nested donations, claims, votes, contributions, adjustments
        '404':
          description: Donor not found
  /admin/donors/{id}/role:
    patch:
      tags: [Admin]
      summary: Set a donor's role (USER, MODERATOR, or ADMIN)
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [USER, MODERATOR, ADMIN] }
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  email: { type: string }
                  role: { type: string, enum: [USER, MODERATOR, ADMIN] }
        '400':
          description: Invalid role
  /admin/donors/{id}/revoke-token:
    post:
      tags: [Admin]
      summary: Revoke donor magic token
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Token revoked
  /admin/donors/{id}/regenerate-token:
    post:
      tags: [Admin]
      summary: Regenerate donor magic token
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: New token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  email: { type: string }
                  magic_token: { type: string }
  /admin/donors/{id}/freeze:
    post:
      tags: [Admin]
      summary: Freeze/unfreeze a donor account
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [frozen]
              properties:
                frozen: { type: boolean }
      responses:
        '200':
          description: Freeze status updated
...
  /admin/donors/{id}/reverse-spend:
    post:
      tags: [Admin]
      summary: Reverse a claim, vote, or contribution (refunds donor)
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [spend_type, spend_id]
              properties:
                spend_type:
                  type: string
                  enum: [claim, vote, contribution]
                spend_id: { type: string }
      responses:
        '200':
          description: Spend reversed
...
  /admin/polls/{id}/options:
    post:
      tags: [Admin]
      summary: Add a poll option
      security: [{ AdminKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label: { type: string }
      responses:
        '200':
          description: Option created
...
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin API key matching `ADMIN_API_KEY` env var